requirements.txt
flask
pandas
mistralai>=1.0.0
python-dotenv
openpyxl
tenacity 